+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 ===
 
+### 19.08.2022
+- Added readme and changelog resource uploads
+- Using log levels FINER and FINE 
+- Added default values to all parameters
+- Return token even on failure if a token is returned upon create a new API
+- Fixed enabling dry run bug
+
 ### 29.06.2022
 - Code added to the jenkins git repo
 - Added jar libraries to local repository